pyproject.toml
[tool.poetry]
name = "tgbot-flask"
version = "0.1.0"
description = ""
authors = ["Viacheus <glorianfox@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "20.7"
python-dotenv = "^1.1.0"
openai = "^1.78.1"
sqlalchemy = "^2.0.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
